docs: 초기 파일 셋팅
