+++ src/App.jsx
@@ -1,0 +1,7 @@
+import './App.css';
+
+function App() {
+  return <div className='App'>init</div>;
+}
+
+export default App;
